+++ packages/router/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.0-alpha.14"></a>
+# 2.0.0-alpha.14 (2021-07-25)
+
+### Refactorings:
+
+* **controller:** rename semi public APIs ([c2ee6e9](https://github.com/aurelia/aurelia/commit/c2ee6e9))
+
 <a name="2.0.0-alpha.13"></a>
 # 2.0.0-alpha.13 (2021-07-19)